+++ retz-client/src/main/java/io/github/retz/cli/CommandSchedule.java
@@ -48,6 +48,8 @@ public class CommandSchedule implements SubCommand {
     private int gpu = 0;
     @Parameter(names = {"-prio", "-priority"}, description = "Job priority")
     private int priority = 0;
+    @Parameter(names = {"-N", "--name"}, description = "Human readable job name")
+    private String name;
 
     @Override
     public String getName() {
@@ -65,6 +67,7 @@ public class CommandSchedule implements SubCommand {
 
         Job job = new Job(appName, remoteCmd, envProps, cpu, mem, gpu, ports);
         job.setPriority(priority);
+        job.setName(name);
 
         try (Client webClient = Client.newBuilder(fileConfig.getUri())
                 .setAuthenticator(fileConfig.getAuthenticator())